YouTube-Transcript-to-Blog-Generator/streamlit-app/requirements.txt
crewai
crewai_tools
python-dotenv
langchain-huggingface
beautifulsoup4
yt-dlp
groq
streamlit